Move navbar component into components directory
